requirements.txt
accelerate==0.27.0
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anykeystore==0.2
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==21.2.0
bidict==0.23.1
blessed==1.20.0
boto3==1.34.113
botocore==1.34.113
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
clip-openai==1.0.post20230121
cmake==3.29.3
cramjam==2.8.3
crcmod==1.7
cryptacular==1.6.2
cryptography==39.0.2
cycler==0.12.1
datasets
diffusers==0.30.1
decorator==5.1.1
decord==0.6.0
deepspeed==0.15.0
defusedxml==0.7.1
Deprecated==1.2.14
descartes==1.1.0
dill==0.3.8
distlib==0.3.8
distro-info==1.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
ecdsa==0.19.0
einops==0.6.0
exceptiongroup==1.2.1
executing==2.0.1
fairscale==0.4.13
fastparquet==2024.5.0
ffmpegcv==0.3.13
filelock==3.14.0
fire==0.6.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.6.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.23.2
hupper==1.12.1
idna==3.7
imageio==2.34.1
imgaug==0.2.6
iniconfig==2.0.0
ipaddress==1.0.23
ipdb==0.13.13
ipython==8.18.1
jaxtyping==0.2.28
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonargparse==4.14.1
jsonlines==4.0.0
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
lightning==2.2.3
lightning-utilities==0.11.2
lit==18.1.6
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
miscreant==0.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.24.4
nuscenes-devkit==1.1.11
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.24.0
openai-clip
opencv-python==4.9.0.80
opencv-python-headless==3.4.18.65
packaging==22.0
pandas==1.5.3
parquet==1.3.1
parso==0.8.4
PasteDeploy==3.1.0
pathlib2==2.3.7.post1
pathtools==0.1.2
pbkdf2==1.3
pexpect==4.9.0
pillow==10.3.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
ply==3.11
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==9.0.0
py-spy==0.3.14
pyarrow==11.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pycocotools==2.0.7
pycparser==2.22
pycryptodomex==3.20.0
pycurl==7.43.0.6
Pygments==2.18.0
PyJWT==2.8.0
pynvml==11.5.0
pyope==0.2.2
pyOpenSSL==23.2.0
pyparsing==3.1.2
pyquaternion==0.9.9
pyramid==2.0.2
pyramid-mailer==0.15.1
pytest==6.2.5
python-consul==1.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-etcd==0.4.5
python-jose==3.3.0
python-socketio==5.11.2
python3-openid==3.2.0
pytorch-extension==0.2
pytorch-lightning==2.2.3
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
schedule==1.2.2
scikit-image==0.22.0
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.2.0
sentry-sdk==2.3.1
setproctitle==1.3.3
Shapely==1.8.5.post1
shortuuid==1.0.13
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.30
stack-data==0.6.3
sympy==1.12
taming-transformers-rom1504==0.0.6
tenacity==8.3.0
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.5.0
thriftpy2==0.5.0
tifffile==2024.5.22
timm==1.0.3
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.3.1
torch-fidelity==0.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.1
tox==3.28.0
tqdm==4.66.4
traitlets==5.14.3
transaction==4.0
transformers==4.41.1
translationstring==1.4
typeguard==2.13.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.18
velruse==1.1.1
venusian==3.1.0
virtualenv==20.26.2
wandb==0.17.0
watchdog==4.0.1
wcwidth==0.2.13
webdataset==0.2.86
WebOb==1.8.7
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xformers==0.0.25
xxhash==3.4.1
yarl==1.9.4
zope.deprecation==5.0
zope.interface==6.4.post2
zope.sqlalchemy==3.1

